[ISSUE #22452] code review
